+++ R/msharpeScreening.R
@@ -21,8 +21,12 @@
   }
 
   # size of inputs and outputs
+  X <- as.matrix(X)
   T <- nrow(X)
   N <- ncol(X)
+  if (N < 2L) {
+    stop("within-group screening needs at least two funds in 'X'; supply 'Y' to screen a single fund against a peer group")
+  }
   pval <- dmsharpe <- tstat <- matrix(data = NA, N, N)
 
   # determine which pairs can be compared (in a matrix way)
@@ -37,6 +41,7 @@
 
   if (length(liststocks) > 1) {
     cl <- parallel::makeCluster(ctr$nCore)
+    on.exit(parallel::stopCluster(cl), add = TRUE)
 
     liststocks <- liststocks[1:(length(liststocks) - 1)]
 
@@ -44,7 +49,6 @@
                                   rdata = X, level = level, T = T, N = N, na.neg = na.neg, nBoot = ctr$nBoot,
                                   bsids = bsids, minObs = ctr$minObs, type = ctr$type, hac = ctr$hac,
                                   b = ctr$bBoot, ttype = ctr$ttype, pBoot = ctr$pBoot)
-    parallel::stopCluster(cl)
 
     for (i in 1:length(liststocks)) {
       out <- z[[i]]